moving stuff to utils
